Fix compilation with strict warnings

+++ engines/uadstu/dstu_cipher.c
@@ -58,7 +58,7 @@ static int dstu_cipher_do_cipher(EVP_CIPHER_CTX *ctx, unsigned char *out,
         return -1;
 
     if (ctx->num) {
-        to_use = (ctx->num < inl) ? ctx->num : inl;
+        to_use = ((size_t)ctx->num < inl) ? (size_t)ctx->num : inl;
 
         for (i = 0; i < to_use; i++) {
             if (ctx->encrypt) {
